Change SecretCode type and create random number generator function

+++ week-09/day-4/URLAliaser/URLAliaser/Models/Entities/Link.cs
@@ -5,6 +5,6 @@ namespace URLAliaser.Models.Entities
         public long Id { get; set; }
         public string LinkURL { get; set; }
         public string Alias { get; set; }
-        public long SecretCode { get; set; }
+        public int SecretCode { get; set; }
     }
 }
